Enum linera::ClientCommand

pub(crate) enum ClientCommand {

    Transfer {
        sender: Account,
        recipient: Account,
        amount: Amount,
    },
    OpenChain {
        sender: Option<ChainId>,
        public_key: Option<PublicKey>,
    },
    CloseChain {
        sender: ChainId,
    },
    QueryBalance {
        chain_id: Option<ChainId>,
    },
    SyncBalance {
        chain_id: Option<ChainId>,
    },
    QueryValidators {
        chain_id: Option<ChainId>,
    },
    SetValidator {
        name: ValidatorName,
        address: String,
        votes: u64,
    },
    RemoveValidator {
        name: ValidatorName,
    },
    Pricing {
        certificate: Option<Amount>,
        fuel: Option<Amount>,
        storage: Option<Amount>,
        messages: Option<Amount>,
    },
    Benchmark {
        max_in_flight: u64,
        max_proposals: Option<usize>,
    },
    CreateGenesisConfig {
        committee_config_path: PathBuf,
        genesis_config_path: PathBuf,
        admin_root: u32,
        initial_funding: Amount,
        start_timestamp: Option<DateTime<Utc>>,
        num: u32,
        certificate_price: Amount,
        fuel_price: Amount,
        storage_price: Amount,
        messages_price: Amount,
    },
    Watch {
        chain_id: Option<ChainId>,
        raw: bool,
    },
    Service {
        chain_id: Option<ChainId>,
        config: ChainListenerConfig,
        port: NonZeroU16,
    },
    PublishBytecode {
        contract: PathBuf,
        service: PathBuf,
        publisher: Option<ChainId>,
    },
    CreateApplication {
        bytecode_id: BytecodeId,
        creator: Option<ChainId>,
        json_parameters: Option<String>,
        json_parameters_path: Option<PathBuf>,
        json_argument: Option<String>,
        json_argument_path: Option<PathBuf>,
        required_application_ids: Option<Vec<UserApplicationId>>,
    },
    PublishAndCreate {
        contract: PathBuf,
        service: PathBuf,
        publisher: Option<ChainId>,
        json_parameters: Option<String>,
        json_parameters_path: Option<PathBuf>,
        json_argument: Option<String>,
        json_argument_path: Option<PathBuf>,
        required_application_ids: Option<Vec<UserApplicationId>>,
    },
    RequestApplication {
        application_id: UserApplicationId,
        target_chain_id: Option<ChainId>,
        requester_chain_id: Option<ChainId>,
    },
    Keygen,
    Assign {
        key: PublicKey,
        message_id: MessageId,
    },
    Wallet(WalletCommand),
    Project(ProjectCommand),

}

Variants

Transfer

Fields

sender: Account

Sending chain id (must be one of our chains)

recipient: Account

Recipient account

amount: Amount

Amount to transfer

Transfer funds

OpenChain

Fields

sender: Option<ChainId>

Sending chain id (must be one of our chains).

public_key: Option<PublicKey>

Public key of the new owner (otherwise create a key pair and remember it)

Open (i.e. activate) a new chain deriving the UID from an existing one.

CloseChain

Fields

sender: ChainId

Sending chain id (must be one of our chains)

Close (i.e. deactivate) an existing chain.

QueryBalance

Fields

chain_id: Option<ChainId>

Chain id

Read the balance of the chain from the local state of the client.

SyncBalance

Fields

chain_id: Option<ChainId>

Chain id

Synchronize the local state of the chain (including a conservative estimation of the available balance) with a quorum validators.

QueryValidators

Fields

chain_id: Option<ChainId>

Chain id

Show the current set of validators for a chain.

SetValidator

Fields

name: ValidatorName

The public key of the validator.

address: String

Network address

votes: u64

Voting power

Add or modify a validator (admin only)

RemoveValidator

Fields

name: ValidatorName

The public key of the validator.

Remove a validator (admin only)

Pricing

Fields

certificate: Option<Amount>

Set the base price for each certificate.

fuel: Option<Amount>

Set the price per unit of fuel when executing user messages and operations.

storage: Option<Amount>

Set the price per byte to store a block’s operations, incoming and outgoing messages.

messages: Option<Amount>

Set the price per byte to store and send outgoing cross-chain messages.

View or update the pricing.

Benchmark

Fields

max_in_flight: u64

Maximum number of blocks in flight

max_proposals: Option<usize>

Use a subset of the chains to generate N transfers

Send one transfer per chain in bulk mode

CreateGenesisConfig

Fields

committee_config_path: PathBuf

Sets the file describing the public configurations of all validators

genesis_config_path: PathBuf

The output config path to be consumed by the server

admin_root: u32

Index of the admin chain in the genesis config

initial_funding: Amount

Known initial balance of the chain

start_timestamp: Option<DateTime<Utc>>

The start timestamp: no blocks can be created before this time.

num: u32

Number of additional chains to create

certificate_price: Amount

Set the base price for each certificate.

fuel_price: Amount

Set the price per unit of fuel when executing user messages and operations.

storage_price: Amount

Set the price per byte to store a block’s operations, incoming and outgoing messages.

messages_price: Amount

Set the price per byte to store and send outgoing cross-chain messages.

Create genesis configuration for a Linera deployment. Create initial user chains and print information to be used for initialization of validator setup. This will also create an initial wallet for the owner of the initial “root” chains.

Watch

Fields

chain_id: Option<ChainId>

The chain id to watch.

raw: bool

Show all notifications from all validators.

Watch the network for notifications.

Service

Fields

chain_id: Option<ChainId>

Chain id

config: ChainListenerConfig

port: NonZeroU16

The port on which to run the server

Run a GraphQL service on the local node of a given chain.

PublishBytecode

Fields

contract: PathBuf

Path to the WASM file for the application “contract” bytecode.

service: PathBuf

Path to the WASM file for the application “service” bytecode.

publisher: Option<ChainId>

An optional chain ID to publish the bytecode. The default chain of the wallet is used otherwise.

Publish bytecode.

CreateApplication

Fields

bytecode_id: BytecodeId

The bytecode ID of the application to create.

creator: Option<ChainId>

An optional chain ID to host the application. The default chain of the wallet is used otherwise.

json_parameters: Option<String>

The shared parameters as JSON string.

json_parameters_path: Option<PathBuf>

Path to a JSON file containing the shared parameters.

json_argument: Option<String>

The initialization argument as a JSON string.

json_argument_path: Option<PathBuf>

Path to a JSON file containing the initialization argument.

required_application_ids: Option<Vec<UserApplicationId>>

The list of required dependencies of application, if any.

Create an application.

PublishAndCreate

Fields

contract: PathBuf

Path to the WASM file for the application “contract” bytecode.

service: PathBuf

Path to the WASM file for the application “service” bytecode.

publisher: Option<ChainId>

An optional chain ID to publish the bytecode. The default chain of the wallet is used otherwise.

json_parameters: Option<String>

The shared parameters as JSON string.

json_parameters_path: Option<PathBuf>

Path to a JSON file containing the shared parameters.

json_argument: Option<String>

The initialization argument as a JSON string.

json_argument_path: Option<PathBuf>

Path to a JSON file containing the initialization argument.

required_application_ids: Option<Vec<UserApplicationId>>

The list of required dependencies of application, if any.

Create an application, and publish the required bytecode.

RequestApplication

Fields

application_id: UserApplicationId

The ID of the application to request.

target_chain_id: Option<ChainId>

The target chain on which the application is already registered. If not specified, the chain on which the application was created is used.

requester_chain_id: Option<ChainId>

The owned chain on which the application is missing.

Request an application from another chain, so it can be used on this one.

Keygen

Create an unassigned key-pair.

Assign

Fields

key: PublicKey

The public key to assign.

message_id: MessageId

The ID of the message that created the chain. (This uniquely describes the chain and where it was created.)

Link a key owned by the wallet to a chain that was just created for that key.

Wallet(WalletCommand)

Show the contents of the wallet.

Project(ProjectCommand)

Manage Linera projects.